The place card lands its minimize without the last moment hop, the results list folds its filter chips away with the sheet, and menu photos show when they were uploaded
The card now holds at the minimized height while the fold is running
instead of chasing its shrinking content those last few pixels, which
is what the small hop at the end was. The results sheet gives its
filter chips the same fold the place card uses, so they shrink away
with the sheet instead of vanishing after it has already stopped. The
fold helper moved to ui/SheetFold.kt since both sheets share it now.
Menu photos carry the upload date we already scrape as a small corner
stamp, so you can tell whether a menu shot is recent without opening
it.
